Update app URL with HTTPS
HTTPS is needed for getUserMedia in chrome. I've finally got it set up
with FREE hosting with FREE SSL on a FREE domain. It's hosted with
pancake.io (with a certificate issued to *.pancakeapps.com, which
doesn't work with a custom domain) through cloudflare, which
(eventually!) gives it a new, valid certificate. I'm using non-strict
Full SSL mode so it ignores the invalidity of the pancake cert and
encrypts it all the way.