Use more styles from elementary
Propagate non-100% black/white colors thru the app more. The more subtle colors now appear in the menus, and waveform in the dark theme. The light theme already used it but it's now referencing rather than copying the color value. Similarly the dark background color is now by reference; however, the light theme doesn't use the background color of  so that's staying as a copied color for now.

Improve the button pressed states by increasing precedence. With the light theme it improves contrast; with the dark theme it's technically worse. But it isn't and wasn't good for either. At least we have higher-contrast themes.